core: move store update into tranfer process manager
